refactor(speech): extract reading for Reader
